refactor PropertyReader
