+++ src/test/java/framework/BaseEntity.java
@@ -1,5 +1,7 @@
 package framework;
 
+import framework.factory.BrowserFactory;
+import framework.factory.DriverManager;
 import org.openqa.selenium.WebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
