+++ src/test/java/framework/elements/BaseElement.java
@@ -1,7 +1,7 @@
 package framework.elements;
 
 import framework.BaseEntity;
-import framework.BrowserFactory;
+import framework.factory.BrowserFactory;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
