+++ src/test/java/framework/factory/PropertyReader.java
@@ -1,4 +1,4 @@
-package framework;
+package framework.factory;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -11,14 +11,7 @@ public class PropertyReader {
     private static final Object lock = new Object();
     private static PropertyReader instance;
 
-    private String url;
-    private String browserName;
-    private long shortTimeOut;
-    private long longTimeOut;
-    private String host;
-    private String sqlLogin;
-    private String sqlPassword;
-
+    private Properties properties;
 
     protected static PropertyReader getInstance() {
         if (instance == null) {
@@ -31,7 +24,7 @@ public class PropertyReader {
     }
 
     private void loadData() {
-        Properties properties = new Properties();
+        properties = new Properties();
         try {
             properties.load(path);
         } catch (FileNotFoundException e) {
@@ -40,29 +33,9 @@ public class PropertyReader {
         } catch (IOException e) {
             System.err.println("Bad canonical path");
         }
-
-        url = properties.getProperty("url");
-        browserName = properties.getProperty("browserName");
-        shortTimeOut = Long.parseLong(properties.getProperty("shortTimeOut"));
-        longTimeOut = Long.parseLong(properties.getProperty("longTimeOut"));
-        host = properties.getProperty("db.host");
-        sqlLogin = properties.getProperty("db.login");
-        sqlPassword = properties.getProperty("db.password");
     }
 
-    public String getUrl() {
-        return url;
-    }
-
-    public String getBrowserName() {
-        return browserName;
-    }
-
-    public long getShortTimeOut() {
-        return shortTimeOut;
-    }
-
-    public long getLongTimeOut() {
-        return longTimeOut;
+    public String getProperty(String property) {
+        return properties.getProperty(property);
     }
 }
